simplify messagehandler code
