+++ messages/teleporterv2/message_handler.go
@@ -6,12 +6,9 @@ package teleporterv2
 import (
 	"context"
 	"fmt"
-	"slices"
-	"time"
 
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/utils/logging"
-	"github.com/ava-labs/avalanchego/utils/set"
 	"github.com/ava-labs/avalanchego/vms/platformvm/warp"
 	avalancheWarp "github.com/ava-labs/avalanchego/vms/platformvm/warp"
 	warpPayload "github.com/ava-labs/avalanchego/vms/platformvm/warp/payload"
@@ -29,7 +26,6 @@ import (
 	ethereum "github.com/ava-labs/libevm"
 	"github.com/ava-labs/libevm/accounts/abi/bind"
 	"github.com/ava-labs/libevm/common"
-	"github.com/ava-labs/libevm/core/types"
 	"go.uber.org/zap"
 )
 
@@ -47,21 +43,14 @@ type factory struct {
 	sourceSubnetID  ids.ID
 }
 
+// messageHandler relays TeleporterV2 messages whose destination messenger verifies them against a
+// MerkleValidatorSetRegistry. The attestation passed to receiveCrossChainMessage is a Merkle
+// attestation over the validator set committed under the registry's stored root, so the signature
+// must be aggregated over the source subnet's set at the committed P-chain height.
 type messageHandler struct {
-	logger              logging.Logger
-	teleporterMessage   *teleportermessengerv2.TeleporterMessageV2
-	unsignedMessage     *warp.UnsignedMessage
-	destinationClient   vms.DestinationClient
-	pChainClient        clients.CanonicalValidatorState
-	sourceSubnetID      ids.ID
-	signatureAggregator *aggregator.SignatureAggregator
-	metrics             messages.Metrics
-	quorumNumerator     uint64
-	// teleporterMessageID identifies the message for delivery/replay checks.
-	teleporterMessageID ids.ID
-	messageConfig       *Config
-	// teleporterAddress is the TeleporterMessengerV2 contract (receive target + message ID + status).
-	teleporterAddress common.Address
+	handlerBase
+	pChainClient   clients.CanonicalValidatorState
+	sourceSubnetID ids.ID
 }
 
 // NewMessageHandlerFactory creates a factory for the TeleporterV2 message protocol. The
@@ -132,35 +121,30 @@ func (f *factory) NewMessageHandler(
 		zap.Stringer("teleporterAddress", f.messageConfig.teleporterAddress()),
 	}
 
-	if f.messageConfig.VerifierType == VerifierTypeWarp {
-		return &warpMessageHandler{
-			logger:              logger.With(logFields...),
-			teleporterMessage:   teleporterMessage,
-			unsignedMessage:     unsignedMessage,
-			destinationClient:   destinationClient,
-			signatureAggregator: signatureAggregator,
-			metrics:             metrics,
-			signingSubnetID:     signingSubnetID,
-			quorumNumerator:     quorumNumerator,
-			teleporterMessageID: teleporterMessageID,
-			messageConfig:       f.messageConfig,
-			teleporterAddress:   f.messageConfig.teleporterAddress(),
-		}, nil
-	}
-
-	return &messageHandler{
+	base := handlerBase{
 		logger:              logger.With(logFields...),
 		teleporterMessage:   teleporterMessage,
 		unsignedMessage:     unsignedMessage,
 		destinationClient:   destinationClient,
-		pChainClient:        f.pChainClient,
-		sourceSubnetID:      f.sourceSubnetID,
 		signatureAggregator: signatureAggregator,
 		metrics:             metrics,
 		quorumNumerator:     quorumNumerator,
 		teleporterMessageID: teleporterMessageID,
 		messageConfig:       f.messageConfig,
 		teleporterAddress:   f.messageConfig.teleporterAddress(),
+	}
+
+	if f.messageConfig.VerifierType == VerifierTypeWarp {
+		return &warpMessageHandler{
+			handlerBase:     base,
+			signingSubnetID: signingSubnetID,
+		}, nil
+	}
+
+	return &messageHandler{
+		handlerBase:    base,
+		pChainClient:   f.pChainClient,
+		sourceSubnetID: f.sourceSubnetID,
 	}, nil
 }
 
@@ -179,111 +163,43 @@ func (f *factory) GetMessageRoutingInfo(
 	}, nil
 }
 
-// ShouldSendMessage returns true if the message should be relayed to the destination chain.
-func (m *messageHandler) ShouldSendMessage() (bool, error) {
-	// RequiredGasLimit is a Solidity uint256 (*big.Int). Calling Uint64() on a value that does not
-	// fit in 64 bits is undefined, so treat any non-uint64 value as exceeding the block gas limit.
-	destBlockGasLimit := m.destinationClient.BlockGasLimit()
-	if !m.teleporterMessage.RequiredGasLimit.IsUint64() ||
-		m.teleporterMessage.RequiredGasLimit.Uint64() > destBlockGasLimit {
-		m.logger.Info(
-			"Gas limit exceeds maximum threshold",
-			zap.Stringer("requiredGasLimit", m.teleporterMessage.RequiredGasLimit),
-			zap.Uint64("blockGasLimit", destBlockGasLimit),
-		)
-		return false, nil
-	}
-
-	if !containsAllowedRelayer(m.teleporterMessage.AllowedRelayerAddresses, m.destinationClient.SenderAddresses()) {
-		m.logger.Info("Relayer EOA not allowed to deliver this message.")
-		return false, nil
-	}
-
-	teleporterMessenger, err := m.getTeleporterMessenger()
-	if err != nil {
-		return false, err
-	}
-	delivered, err := teleporterMessenger.MessageReceived(&bind.CallOpts{}, m.teleporterMessageID)
-	if err != nil {
-		m.logger.Error(
-			"Failed to check if message has been delivered to destination chain.",
-			zap.Error(err),
-		)
-		return false, err
-	}
-	if delivered {
-		m.logger.Info("Message already delivered to destination.")
-		return false, nil
-	}
-
-	return true, nil
-}
-
 // ProcessMessage relays the message to the destination chain by aggregating a signature over the
 // committed validator set and delivering it via SendMessage. It does not retry on failure or
 // checkpoint the height. Returns the transaction hash if the message is successfully relayed.
 func (m *messageHandler) ProcessMessage() (common.Hash, error) {
-	m.logger.Info("Relaying message")
-	shouldSend, err := m.ShouldSendMessage()
-	if err != nil {
-		m.metrics.IncFailedRelayMessageCount("failed to check if message should be sent")
-		return common.Hash{}, fmt.Errorf("failed to check if message should be sent: %w", err)
-	}
-	if !shouldSend {
-		m.logger.Info("Message should not be sent")
-		return common.Hash{}, nil
-	}
+	return m.relay(func(ctx context.Context) (common.Hash, error) {
+		sourceChainID := m.unsignedMessage.SourceChainID
 
-	ctx, cancel := context.WithTimeout(context.Background(), utils.DefaultCreateSignedMessageTimeout)
-	defer cancel()
+		// Fetch the validator set committed under the registry's stored Merkle root. The signature
+		// must be aggregated over the exact set (and P-chain height) the root was built from, so the
+		// signer bitset and weights match the committed total and the leaves resolve against the
+		// stored root.
+		commitment, err := m.fetchCommitment(ctx, sourceChainID)
+		if err != nil {
+			m.metrics.IncFailedRelayMessageCount("failed to read committed validator set")
+			m.logger.Error("Failed to read committed validator set", zap.Error(err))
+			return common.Hash{}, err
+		}
 
-	sourceChainID := m.unsignedMessage.SourceChainID
+		validators, err := m.validatorsAtCommitment(ctx, commitment)
+		if err != nil {
+			m.metrics.IncFailedRelayMessageCount("failed to fetch committed validators")
+			m.logger.Error("Failed to fetch committed validator set", zap.Error(err))
+			return common.Hash{}, err
+		}
 
-	// Fetch the validator set committed under the registry's stored Merkle root. The signature must
-	// be aggregated over the exact set (and P-chain height) the root was built from, so the signer
-	// bitset and weights match the committed total and the leaves resolve against the stored root.
-	commitment, err := m.fetchCommitment(ctx, sourceChainID)
-	if err != nil {
-		m.metrics.IncFailedRelayMessageCount("failed to read committed validator set")
-		m.logger.Error("Failed to read committed validator set", zap.Error(err))
-		return common.Hash{}, err
-	}
+		signedMessage, err := m.signMessage(ctx, m.sourceSubnetID, commitment.PChainHeight)
+		if err != nil {
+			return common.Hash{}, err
+		}
 
-	validators, err := m.validatorsAtCommitment(ctx, commitment)
-	if err != nil {
-		m.metrics.IncFailedRelayMessageCount("failed to fetch committed validators")
-		m.logger.Error("Failed to fetch committed validator set", zap.Error(err))
-		return common.Hash{}, err
-	}
-
-	startCreateSignedMessageTime := time.Now()
-	signedMessage, err := m.signatureAggregator.CreateSignedMessage(
-		ctx,
-		m.logger,
-		m.unsignedMessage,
-		nil,
-		m.sourceSubnetID,
-		m.quorumNumerator,
-		commitment.PChainHeight,
-	)
-	m.metrics.IncFetchSignatureAppRequestCount()
-	if err != nil {
-		m.metrics.IncFailedRelayMessageCount("failed to create signed warp message via AppRequest network")
-		return common.Hash{}, fmt.Errorf("failed to create signed warp message via AppRequest network: %w", err)
-	}
-	m.metrics.SetCreateSignedMessageLatencyMS(float64(time.Since(startCreateSignedMessageTime).Milliseconds()))
-
-	txHash, err := m.SendMessage(ctx, signedMessage, validators)
-	if err != nil {
-		m.metrics.IncFailedRelayMessageCount("failed to send warp message")
-		return common.Hash{}, fmt.Errorf("failed to send warp message: %w", err)
-	}
-	m.logger.Info(
-		"Finished relaying message to destination chain",
-		zap.Stringer("txID", txHash),
-	)
-	m.metrics.IncSuccessfulRelayMessageCount()
-	return txHash, nil
+		txHash, err := m.SendMessage(ctx, signedMessage, validators)
+		if err != nil {
+			m.metrics.IncFailedRelayMessageCount("failed to send warp message")
+			return common.Hash{}, fmt.Errorf("failed to send warp message: %w", err)
+		}
+		return txHash, nil
+	})
 }
 
 // SendMessage builds a Merkle attestation for the signed message and delivers it to the
@@ -294,8 +210,6 @@ func (m *messageHandler) SendMessage(
 	signedMessage *warp.Message,
 	validators []*validatorupdater.Validator,
 ) (common.Hash, error) {
-	sourceChainID := m.unsignedMessage.SourceChainID
-
 	bitSetSig, ok := signedMessage.Signature.(*avalancheWarp.BitSetSignature)
 	if !ok {
 		return common.Hash{}, fmt.Errorf("expected BitSetSignature, got %T", signedMessage.Signature)
@@ -310,7 +224,7 @@ func (m *messageHandler) SendMessage(
 	callData, err := teleportermessengerv2.PackReceiveCrossChainMessageMerkle(
 		*m.teleporterMessage,
 		m.unsignedMessage.NetworkID,
-		sourceChainID,
+		m.unsignedMessage.SourceChainID,
 		attestation.Bytes(),
 		m.messageConfig.rewardAddress(),
 	)
@@ -325,46 +239,8 @@ func (m *messageHandler) SendMessage(
 		return common.Hash{}, err
 	}
 
-	receipt, err := m.destinationClient.SendTx(
-		m.logger,
-		nil, // No access list: verification reads the attestation from calldata, not a predicate.
-		set.Of(m.teleporterMessage.AllowedRelayerAddresses...),
-		m.teleporterAddress,
-		gasLimit,
-		callData,
-	)
-	if err != nil {
-		m.logger.Error("Failed to send tx.", zap.Error(err))
-		return common.Hash{}, err
-	}
-
-	txHash := receipt.TxHash
-	log := m.logger.With(zap.Stringer("txID", txHash))
-	if receipt.Status == types.ReceiptStatusSuccessful {
-		log.Info("Delivered message to destination chain")
-		return txHash, nil
-	}
-
-	// The transaction reverted. A common benign cause is that the message was already delivered
-	// by another relayer, so check delivery status before treating the revert as a failure.
-	teleporterMessenger, err := m.getTeleporterMessenger()
-	if err != nil {
-		log.Error("Transaction failed", zap.Error(err))
-		return common.Hash{}, fmt.Errorf("transaction failed with status: %d", receipt.Status)
-	}
-
-	delivered, err := teleporterMessenger.MessageReceived(&bind.CallOpts{}, m.teleporterMessageID)
-	if err != nil {
-		log.Error("Transaction failed", zap.Error(err))
-		return common.Hash{}, fmt.Errorf("transaction failed with status: %d", receipt.Status)
-	}
-	if delivered {
-		log.Info("Execution reverted: message already delivered to destination.")
-		return txHash, nil
-	}
-
-	log.Error("Transaction failed")
-	return common.Hash{}, fmt.Errorf("transaction failed with status: %d", receipt.Status)
+	// No access list: verification reads the attestation from calldata, not a predicate.
+	return m.sendTxAndConfirm(nil, gasLimit, callData)
 }
 
 // fetchCommitment reads the registry's stored validator set commitment for the source chain,
@@ -460,14 +336,6 @@ func (m *messageHandler) selectSenderAddress() common.Address {
 	return common.Address{}
 }
 
-func (m *messageHandler) getTeleporterMessenger() (*teleportermessengerv2.TeleporterMessengerV2, error) {
-	messenger, err := teleportermessengerv2.NewTeleporterMessengerV2(m.teleporterAddress, m.destinationClient.Client())
-	if err != nil {
-		return nil, fmt.Errorf("failed to bind teleporter v2 messenger: %w", err)
-	}
-	return messenger, nil
-}
-
 // parseMessage decodes the TeleporterV2 message that [message] was sent as. Both adapters send
 // their messages through the Warp precompile, so the source message payload is the encoded
 // unsigned Warp message that carries the Teleporter message as its addressed payload. The two
@@ -498,20 +366,4 @@ func (f *factory) parseMessage(
 		}
 	}
 	return unsignedMessage, teleporterMessage, nil
-}
-
-func isAllowedRelayer(allowedRelayers []common.Address, eoa common.Address) bool {
-	if len(allowedRelayers) == 0 {
-		return true
-	}
-	return slices.Contains(allowedRelayers, eoa)
-}
-
-func containsAllowedRelayer(allowedRelayers []common.Address, eoas []common.Address) bool {
-	if len(allowedRelayers) == 0 {
-		return true
-	}
-	return slices.ContainsFunc(eoas, func(eoa common.Address) bool {
-		return slices.Contains(allowedRelayers, eoa)
-	})
 }
